requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
langchain-core==0.1.47
langchain-openai==0.1.4
python-dotenv==1.0.1